requirements.txt
# Core dependencies
flask==2.3.3
python-dotenv==1.0.0
flask-sqlalchemy==3.1.1
Werkzeug==2.3.7

# Database drivers
# Uncomment the one you need
psycopg2-binary==2.9.7  # PostgreSQL driver
# mysqlclient==2.2.0      # MySQL driver

# Authentication
PyJWT==2.8.0

# API and serialization
flask-marshmallow==0.15.0
marshmallow-sqlalchemy==0.29.0

# CORS support
flask-cors==4.0.0

# Testing
pytest==7.4.2
pytest-flask==1.2.0

# Development tools
flask-migrate==4.0.5  # For database migrations